scheduler/requirements.txt
# Azure Functions scheduler dependencies
azure-functions
requests
python-dotenv
